requirements.txt
streamlit==1.37.1
opencv-python==4.10.0.84
numpy==2.0.1
pandas==2.2.2
rdflib==7.0.0
pyyaml==6.0.2
Pillow==10.4.0
# Optional (heavier):
ultralytics==8.2.82
torch>=2.2.0
# Optional pose:
mediapipe==0.10.14
